removes debug output from PopplerRenderer
In 5c83fde4c4f252baa5beeed4aa583ba131b88e90 we added debug output from worker threads, because it was observed that some worker thread remains in an infinite loop. Now, that thread has been identified as the one from the PackageAnalyzer class. Particularly, the bug occured when a package couldn't be found (it's not sure whether this is a necessary for the bug to occur). Hence, this commit removes the debug output from the other worker threads.